[Task] Refactor file structure
