+++ src/pages/Home.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import MainHeadline from './components/MainHeadline';
 import ContactUs from './components/ContactUs';
-import FeatureSection from './components/feature/FeatureSection';
-import Testimonial from './components/testimonial/Testimonial';
+import FeatureSection from './components/feature';
+import Testimonial from './components/testimonial';
 
 const Home = () => {
   return (
